+++ packages/subgraph/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the `MembershipContractAnnounced` event.
+- Added `voteReplaced` and `updatedAt` to `TokenVotingVote` and `AddresslistVotingVote`.
 
 ### Changed